Refactored project browser pages.
