+++ lib/pages/projects/project_browser_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:portfolio_website/pages/projects/all_projects_page.dart';
-import 'package:portfolio_website/pages/projects/developments_projects_page.dart';
-import 'package:portfolio_website/pages/projects/ml_projects_page.dart';
+import 'package:portfolio_website/pages/projects/browser_pages/all_projects_page.dart';
+import 'package:portfolio_website/pages/projects/browser_pages/developments_projects_page.dart';
+import 'package:portfolio_website/pages/projects/browser_pages/ml_projects_page.dart';
 import 'package:portfolio_website/providers/project_slider_provider.dart';
 import 'package:provider/provider.dart';
